Updates to current API state.
This is the current API state based on debugging from my
car/account.

Cabin status is gone, API returns a 404 for some reason.

Location is gone for data protect reasons and removed from the
app. Checked the URL in all regions.

Removed some battery status information as it is now returned
blank. This may just be an issue for my car/model.

Fixed battery percentage calculation.

Forked, I intend to maintain this fork and the homekit daemon.