Fix units on monthly report
For the individual trips, the code was assuming that "ElectricMileage"
was in kWh/km but for me it seems to be kWh/mile. The json record provides a
field "ElectricCostScale" which for me is "kWh/mile" so assume
this specifies the units of "ElectricMileage".

For the per-day summary, it had the opposite problem - it calculated
the per-day summary in display units, but then assumed this matches
ms.EfficiencyScale units.